Update the UI to look more notes-like (#67)
* Various UI Tweaks

* moving stuff around

* checkpoint

* Now using the new classes

* Pull out BookAction

* Slight refactoring of the menu logic

* Reorganize menus into an extension

* Update CHANGELOG.md
