+++ GrailDiary/CGFloat+Extensions.swift
@@ -1,0 +1,10 @@
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
+
+import UIKit
+
+public extension CGFloat {
+  func roundedToScreenScale(_ rule: FloatingPointRoundingRule = .toNearestOrAwayFromZero) -> CGFloat {
+    let scale: CGFloat = 1.0 / UIScreen.main.scale
+    return scale * (self / scale).rounded(rule)
+  }
+}
